requirements.txt
torch
transformers
diffusers
timm
numpy<2.0
Pillow
requests
Flask
rembg[cpu]
pybase64